+++ src/frontend/src/containers/AllSearchComponents.tsx
@@ -5,34 +5,23 @@ import { Record } from '../redux/types';
 import {
   loadSearchRecordsMock,
   loadSearchRecords
-} from '../redux/search-results/actions';
+} from '../redux/records/actions';
 
 type Props = {
-  loadSearchRecords: Function;
-  loadSearchRecordsMock: Function;
+  fetch: Function;
   records: Record[];
 };
 
 class AllSearchResults extends Component<Props> {
   componentDidMount() {
-    //switch to loadSearchRecordsAction when go to production
-    // this.props.dispatch(loadSearchRecordsMockAction())
-    // loadSearchRecordsMockActionNotThunk(this.props.dispatch)
-    this.props.loadSearchRecordsMock();
+    this.props.fetch();
+    //When we render the search component, fetch should be passed down to it as a prop.
   }
   render() {
-    //actions vs action creators
-    //NOTE 1*****
-    //Now that the records array available in props, you can render
-    //the SearchResultsComponent and pass it the records as a prop
-    //<SearchResultsComponent records={this.props.records}>
-    //Currently you can now hit the rendered button and console log the records
-    //that were fetched from redux
-    //NOTE 2****
-    //The RecordSearch Component should be imported and rendered here.
-    //fetch should be passed down to the RecordSearchComponent and invoked
-    //when a search has been performed.
-    console.log('props', this.props);
+    // ***NOTE***
+    // Eventually the search and searchResults component will replace the button that is currently rendered.
+    // When this happens fetch will be passed down  to the search componen as a prop and invoked when user searches for records.
+    // The records will then be passed to Redux and passed down as props to searchResults component.
     return (
       <button
         onClick={() => {
@@ -45,16 +34,14 @@ class AllSearchResults extends Component<Props> {
   }
 }
 
-const mapDispatchToProps = {
-  loadSearchRecordsMock,
-  loadSearchRecords
-};
-
 const mapStateToProps = (state: AppState) => ({
   records: state.searchRecordsReducer.records
 });
 
 export default connect(
   mapStateToProps,
-  mapDispatchToProps
+  {
+    //switch fetch: loadSearchRecordsMock to fetch: loadSearchRecords when connect to back end
+    fetch: loadSearchRecordsMock
+  }
 )(AllSearchResults);
