+++ src/frontend/src/redux/records/actions.ts
@@ -1,10 +1,47 @@
-import { Record, LOAD_RECORDS } from '../types';
+import { LOAD_SEARCH_RECORDS } from '../types';
+import getSearchRecords from '../../service/get-records';
 
-// This function is an action creator.
-export function loadRecords(records: Record[]) {
-  // Values returned here are actions.
-  return {
-    type: LOAD_RECORDS,
-    records
-  };
-}
+const fakeRecord = [
+  {
+    name: 'WOODS, LAVELLE D',
+    birth_year: 1970,
+    case_number: 'ZA0081909',
+    citation_number: 'ZA0081909',
+    location: 'Multnomah',
+    date: '07/02/2013',
+    violation_type: 'Offense Violation',
+    current_status: 'Inactive',
+    balance_due: '0',
+    charges: ['dui', 'dwi'],
+    case_detail_link:
+      'https://publicaccess.courts.oregon.gov/PublicAccessLogin/CaseDetail.aspx?CaseID=9036658'
+  }
+];
+
+export const loadSearchRecords = () => (dispatch: Function) => {
+  return getSearchRecords()
+    .then(results => {
+      dispatch({
+        type: LOAD_SEARCH_RECORDS,
+        search_records: results.data
+      });
+    })
+    .catch(error => {
+      dispatch({
+        type: LOAD_SEARCH_RECORDS,
+        search_records: [{ error: error.message }]
+      });
+    });
+};
+
+//make sure that we dont have somethign redundant
+export const loadSearchRecordsMock = () => (dispatch: Function) => {
+  return Promise.resolve(fakeRecord).then(payload => {
+    dispatch({
+      type: LOAD_SEARCH_RECORDS,
+      search_records: payload
+    });
+  });
+};
+
+//move mock on own place
