+++ src/frontend/src/redux/store.tsx
@@ -5,13 +5,13 @@ import { configureStore } from 'redux-starter-kit';
 import { combineReducers } from 'redux';
 
 // Reducers:
-import { recordsReducer } from './records/reducer';
+
 import { systemReducer } from './system/reducer';
-import { searchRecordsReducer } from './search-results/reducer';
+import { searchRecordsReducer } from './records/reducer';
 
 const rootReducer = combineReducers({
   system: systemReducer,
-  records: recordsReducer,
+
   searchRecordsReducer
 });
